# swagger.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: A REST API for managing todos
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Register a new user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
        '409':
          description: Email already exists
  /api/auth/login:
    post:
      summary: Login user
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT authentication token
        '401':
          description: Invalid credentials
  /api/todos:
    get:
      summary: Get all todos
      description: Retrieves a list of all todos for the authenticated user
      responses:
        '200':
          description: A list of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The todo ID
                    title:
                      type: string
                      description: The todo title
                    completed:
                      type: boolean
                      description: Whether the todo is completed
                    createdAt:
                      type: string
                      format: date-time
                      description: Creation timestamp
        '401':
          description: Unauthorized - User not authenticated
    post:
      summary: Create a new todo
      description: Creates a new todo for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: The todo title
                completed:
                  type: boolean
                  description: Whether the todo is completed
                  default: false
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The todo ID
                  title:
                    type: string
                    description: The todo title
                  completed:
                    type: boolean
                    description: Whether the todo is completed
                  createdAt:
                    type: string
                    format: date-time
                    description: Creation timestamp
        '401':
          description: Unauthorized - User not authenticated
        '400':
          description: Bad Request - Invalid input